change zscverifyaddress to 100
